fix: Set new values when change parent recrod.
